scraper/requirements.txt
beautifulsoup4>=4.12.0
PyYAML>=6.0
rapidfuzz>=3.0.0
lxml>=4.9.0
